refactor: Calculator 객체의 책임을 Ranking 객체로 분산
